pyproject.toml
[project]
name = "ai-log"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.1",
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
